Move global types to their own file, begin building out card container UI
